Refactor: 문제 수집 비동기 처리

+++ backend/src/main/java/com/algopulza/backend/api/response/SolvedAcProblemRes.java
@@ -19,7 +19,7 @@ public class SolvedAcProblemRes {
 
     private Integer acceptedUserCount;
 
-    private Integer level;
+    private Long level;
 
     private Integer votedUserCount;
